test: move related_request_id test to test_server.py
Move test_report_progress_passes_related_request_id from
test_176_progress_token.py (which tests falsy progress tokens)
to test_server.py where Context behavior tests belong.
