+++ tests/server/mcpserver/test_server.py
@@ -1,7 +1,7 @@
 import base64
 from pathlib import Path
 from typing import Any
-from unittest.mock import patch
+from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 from inline_snapshot import snapshot
@@ -10,6 +10,8 @@ from starlette.applications import Starlette
 from starlette.routing import Mount, Route
 
 from mcp.client import Client
+from mcp.server.context import ServerRequestContext
+from mcp.server.experimental.request_context import Experimental
 from mcp.server.mcpserver import Context, MCPServer
 from mcp.server.mcpserver.exceptions import ToolError
 from mcp.server.mcpserver.prompts.base import Message, UserMessage
@@ -1415,3 +1417,34 @@ def test_streamable_http_no_redirect() -> None:
 
     # Verify path values
     assert streamable_routes[0].path == "/mcp", "Streamable route path should be /mcp"
+
+
+async def test_report_progress_passes_related_request_id():
+    """Test that report_progress passes the request_id as related_request_id.
+
+    Without related_request_id, the streamable HTTP transport cannot route
+    progress notifications to the correct SSE stream, causing them to be
+    silently dropped. See #953 and #2001.
+    """
+    mock_session = AsyncMock()
+    mock_session.send_progress_notification = AsyncMock()
+
+    request_context = ServerRequestContext(
+        request_id="req-abc-123",
+        session=mock_session,
+        meta={"progress_token": "tok-1"},
+        lifespan_context=None,
+        experimental=Experimental(),
+    )
+
+    ctx = Context(request_context=request_context, mcp_server=MagicMock())
+
+    await ctx.report_progress(50, 100, message="halfway")
+
+    mock_session.send_progress_notification.assert_awaited_once_with(
+        progress_token="tok-1",
+        progress=50,
+        total=100,
+        message="halfway",
+        related_request_id="req-abc-123",
+    )
